python-interest-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
beanie==1.26.0
motor==3.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
